# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vidhai ERP API — multi-site mushroom farming production orchestration
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
...
  /traceability/links:
    get:
      operationId: listBatchLinks
      tags: [traceability]
      responses:
        "200":
          description: All batch links
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BatchLink" }
    post:
      operationId: createBatchLink
      tags: [traceability]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BatchLinkInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BatchLink" }
  /traceability/trace/{ootyGrowingBatchId}:
    get:
      operationId: traceGrowingBatch
      tags: [traceability]
      parameters:
        - { name: ootyGrowingBatchId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Full traceability chain
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TraceabilityChain" }

  # ── SCHEDULING ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, role, locationScope]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        role:
          type: string
          enum: [admin, manager, operator, viewer]
        locationScope:
          type: array
          items: { type: string }
        createdAt: { type: string }
    UserInput:
      type: object
      required: [username, password, displayName, role, locationScope]
      properties:
        username: { type: string }
        password: { type: string }
        displayName: { type: string }
        role:
          type: string
          enum: [admin, manager, operator, viewer]
        locationScope:
          type: array
          items: { type: string }
    UserUpdate:
      type: object
      properties:
        displayName: { type: string }
        role:
          type: string
          enum: [admin, manager, operator, viewer]
        locationScope:
          type: array
          items: { type: string }
        password: { type: string }
...
    BatchLink:
      type: object
      required: [id, createdAt]
      properties:
        id: { type: integer }
        ootyGrowingBatchId: { type: ["integer", "null"] }
        annurBatchId: { type: ["integer", "null"] }
        coimBatchId: { type: ["integer", "null"] }
        labSpawnOutputId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    BatchLinkInput:
      type: object
      properties:
        ootyGrowingBatchId: { type: ["integer", "null"] }
        annurBatchId: { type: ["integer", "null"] }
        coimBatchId: { type: ["integer", "null"] }
        labSpawnOutputId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
    TraceabilityChain:
      type: object
      required: [ootyGrowingBatch, links]
      properties:
        ootyGrowingBatch: { type: object }
        annurBatch: { type: object, nullable: true }
        coimBatch: { type: object, nullable: true }
        labSpawnOutputs:
          type: array
          items: { type: object }
        links:
          type: array
          items: { $ref: "#/components/schemas/BatchLink" }

    # ── SCHEDULING ───────────────────────────────────────────────────────────────
...